Modular Test admin integration file
